requirements.txt
asgiref==3.11.0
contourpy==1.3.3
cycler==0.12.1
Django==5.1.6
fonttools==4.61.1
joblib==1.5.3
kiwisolver==1.4.9
lightgbm==4.6.0
matplotlib==3.10.8
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.4
threadpoolctl==3.6.0
tzdata==2025.2
xgboost==3.1.2
djangorestframework-simplejwt
djangorestframework 
dotenv
psycopg2-binary
drf-spectacular
drf-spectacular-sidecar
